refactor(filesystem): move archiving generic functions/utils to `antarest.core.utils` modules
